Fixed misalignment of RLE4 encoded pixel data
It seems if the length of a line is an odd number, the absolute drawing instruction strangely does not contains the actual nibbles but simply the number of bytes * 2. So for images with with an odd resolution this has resulted in reading the padding nibble as well which would result in a column index one larger than the actual row length. As solution befor the second nibble is written the x position is checked against the width of the image to ensure we stay in bounds.