Update GIFLoader.cs
Fixed an issue in the BitStreamReader so it now correctly removes padding bytes.
The IGIFRenderingBlock interface has been extended to include a Dispose method. Disposing of a frame should be done before the next frame is drawn.

Even though it does not follow the specification, it now sets the background color as transparent if the first frame has a transparent color. This will effectively ignore the set background color from the global color palette.